agent.py
import psycopg2
from google_calendar import get_credentials, build

def get_chat_logs():
    """
    Connects to the database and retrieves all chat logs.
    """
    conn = psycopg2.connect(
        dbname="customer_data",
        user="user",
        password="password",
        host="localhost"
    )
    cur = conn.cursor()
    cur.execute("SELECT customer_name, chat_text FROM chat_logs")
    chat_logs = cur.fetchall()
    cur.close()
    conn.close()
    return chat_logs

def get_calendar_events():
    """
    Connects to the Google Calendar API and retrieves all events.
    """
    creds = get_credentials()
    service = build("calendar", "v3", credentials=creds)
    events_result = service.events().list(calendarId='primary', maxResults=2500).execute()
    return events_result.get('items', [])

def main():
    """
    Main function to run the agent.
    """
    chat_logs = get_chat_logs()
    calendar_events = get_calendar_events()

    # This is a placeholder for the AI logic to compare the chat logs and calendar events.
    # You would typically use a library like spaCy or a custom model to extract entities
    # from the chat text and compare them with the calendar events.
    print("Chat Logs:")
    for log in chat_logs:
        print(log)

    print("\nCalendar Events:")
    for event in calendar_events:
        print(event['summary'])

if __name__ == "__main__":
    main()
